1) Space issue fixing on X Xs XsMax and XR 2) Add close on tap gesture
